MXS-3896: Only read first line for password
Reading the whole input would cause both the password to be wrong as well
as discard all input to the command. As the input can only be iterated
once, the reading of the password should only extract as many characters
as needed. An alternative method would be to read the whole input and
treat the first row as the password and the rest as the commands.

This commit also fixes the way the global options are passed to commands
read from the standard input. Previously only the default values were used
which caused the --password and --user options to be ignored if passed to
the actual maxctrl command. Now the options are correctly passed to each
command read from the standard input.