MXS-3907: Fix enabled capability bits
Xpand doesn't support DEPRECATE_EOF so we shouldn't indicate this to the
client. Any versions before 10.6 also do not support metadata caching.

This commit also solves the bug where the enabled features were decided by
at least one server being present that meets the requirement. Now the
correct thing is done and the features are enabled only if all of the
nodes meet the requirements.