MXS-3760 Reduce unit tests running time for QueryPage
Vuex's states should be stubbed instead of calling actions or mutations
as it slows down the process of running unit tests. In addition,
states in the vuex store won't be reset which may fail other tests.

This stubs states by providing initial value for a particular state
via computed property.