MXS-3909: Disable proxy for localhost
If the http_proxy environment variable is defined, axios will use it. If
only the http_proxy is defined without no_proxy, the maxctrl commands will
fail in very confusing ways. Disabling the HTTP proxy for localhost
requests should prevent this while still allowing it to work with remote
connections.