MXS-3902 Sessions can wait for a connection to become available
When both server max_connection is set and pre.emptive pooling
is on, a session can wait for a connection. This happens whenever
max_connections is reached.

There is no limit on the wait time. This feature is meant for
situations where a connection will always be available shortly.

Sessions which are waiting for a connection to become available
cannot be moved to another RoutingWorker.