MXS-3900: Dump stacktrace via GDB
This allows stacktraces from all threads to be dumped instead of only the
thread that caught the signal. This should help resolve cases where the
SystemD watchdog kills MaxScale and the thread that catches the signal
isn't the one that is stuck.

If GDB isn't installed, the old stacktrace generation code is used. This
helps catch unexpected problems on live production systems while still
allowing an improved version to be generated with minimal effort.