MXS-3917: Pool only valid connections
Before version 6, a connection would only be put into the pool when it was
closed if it was known that it was closed in a controlled manner i.e. a
COM_QUIT packet was routed through it. The current code does not check for
this which allows DCBs that were closed in handleError to be placed into
the pool.

As a result of this behavior, a handful of debug assertions in
MariaDBBackendConnection::do_handle_error fail as they assume the DCB has
not been detached from its owning session. An additional mechanism that
would tell whether a DCB is still valid would need to be added to fix this
in a way that would allow putting connections into the pool that weren't
closed in a controlled manner.

A simple method like moving the assignment of the m_hangup flag before the
actual event is handled wouldn't work as non-network errors also must
handled e.g. a rejection of the authentication should prevent a DCB from
being put into the pool.