MXS-3897 Fix bug that caused thread to become unresponsive
In the avro json send loop a full block was always sent. If the
last block was larger than AVRO_DATA_BURST_SIZE the function
incorrectly returned that there is more to send, which lead
to a loop queuing client_callback() for execution.

Modify code to not cause the loop and respect the
AVRO_DATA_BURST_SIZE, i.e do not send a full block
as the block size can be very large, instead check
the limit as each row is sent.