Fix debug assertion with implicit transactions
When autocommit is disabled, the start of a transaction can begin with a
session command. At this point, the transaction did not yet have a target.
This caused a debug assertion to be hit in the response processing code
that assumes that each transaction has a target and it is set when the
first query of the transaction is routed.

Storing the replier of the session command, i.e. m_sescmd_replier, into
the Trx object solves this problem.