MXS-3760 Fix `worksheet` for not showing `txt-editor-container`
`txtEditorDim.width` is 0 because `txt-editor-container` component
is not rendered completely yet.

`handleSetSidebarPct` needs to be called in `doubleRAF` to
ensure all components are completely rendered. So when
`sidebarPct` value changes, it calls `handleRecalPanesDim`
to get accurate value for `txtEditorDim`.

This also removes the test for checking the call of `handleSetSidebarPct`
in `created` hook as it's not testable since the use of double RAF.