MXS-3927: Set current DCB for delayed queries
When the routing of a query is delayed by session_delay_routing, the
re-routing is done in a scope where the current DCB is not set. This
causes any messages logged at this time to not include the session ID as
the current DCB being processed is used for that. To fix this, the current
DCB can be updated for the duration of the routeQuery call.