MXS-3925 Add grants to created users
The user management is somewhat fragile. Depending on the grants
of the currently logged in user, the creation of the user may fail,
or it may succeed but setting the grants may fail, or the deletion
of the created user, once it is noticed that setting the grants
failed, may fail.