Disable ptrace restriction in signal handlers
We don't want the ptrace restrictions to be off all the time and disabling
them for the duration of the stacktrace generation should be the next best
thing. Ideally, we'd fork a child process, add it to the list of PIDs that
are allowed to trace the main process and use that. Unfortunately, that's
a whole lot more work than it is to disable it temporarily.