MXS-3934: Keep track of volatile Xpand nodes
The list of volatile servers created by xpandmon are stored in a vector
which is exposed by the new Monitor::real_servers() function. This also
solves the problem of having separate code for dynamic and static monitors
in 6: removing the check for the monitor type makes both types of monitors
behave the same way.