MXS-3925 Reject client digestion
Basically either the client or the server can digest the password.
If the client digested the password, then we could use the digested
password as the MariaDB password, which would mean that the actual
NoSQL password would not be stored on the MaxScale host. However,
since the MariaDB password really is the important one, it would not
add much value. Furthermore, a client digested password is not
supported with SCRAM-SHA-256, which is the default mechanism (not
supported yet), so we just won't bother.