MXS-3939: Close backend before retrying on deadlock
Now that session commands are correctly retried, it is possible that a
session command causes an ignorable error to be generated. This is common
with Xpand which can generate such an error for almost any query.

Leaving the connection open when this happens is OK from the connection's
point of view but it causes the backend to think that a session command
with the same ID is executed twice. To prevent this, the connection should
be closed before the transaction is retried. If the connection is left
open, the debug assertion in compare_responses() would be hit.