MXS-3939: Store only latest response
If more than one response is received, only the latest is of
interest. Even with this, it is possible that a response is checked
multiple times.

It is also possible that an older response that is re-executed more than
once starts to differ from the accepted answer. This means that the debug
assertion in compare_responses() is now wrong as it is OK to have a
callback installed for a yet-to-be-accepted session commmand while an
older session command fails.