Retry transactions that end with a mismatch
This allows readwritesplit to be configured to trigger a transaction
replay when a checksum mismatch is seen. This will allow transactions that
fail to replay due to transient result errors to be successfully
replayed. Since the normal transaction replay limits are still in place,
this won't affect the total time a replay can take.