Don't discard connections in transactions
The current master connection should not be discarded when a transaction
is open even if the server appears to not be usable. This must be done as
the discarding process doesn't trigger anything to be retried and relying
on the code in handleError to take care of that is better.

The code in question was also partially dead code as the master could no
longer be set into maintenance mode while it was still labeled as master.