Don't handle errors for replication clients
If the client connection executes any of the commands that replication
clients execute, there's no point in any of the more advanced error
handling as the client can tolerate broken connections.

If the normal handling would be done, a debug assertion in
retry_master_query() would be hit as the code would expect there to be a
current query when there won't be one if at least one response is read
from the server. The result processing for replication connections makes
them behave like commands that have never started. This caused it to slip
past the code that would normally handle it for all other types of
commands.