MXS-3934 Add servers to book-keeping only if not added already
A service that has a monitor as its cluster will get its servers
added by the monitor. As a monitor can be stopped and started
again, it must be prevented that the monitor at that point adds
the same server again to the service.

This can be accomplished by checking whether the server already
is present in the m_cluster_servers vector before adding it to
that vector and to the services. That works, because the
vector is never emptied but only added to. A server that
disappears will have its status set to 0, which will cause it
to be ignored.