MXS-3938 Do not assert on found volatile server
As the volatile servers are not deactivated when the xpand
monitor is stopped, they may be present when it is started.
Consequently, even though the xpand monitor is not explictly
aware of a volatile server and thinks it has to be created, the
server may still be found by name.