MXS-3943: Check monitors specifications during alter
If the monitor used the new configuration system, the new parameters were
rejected as unknown parameters. For the next major release, the monitors
should be moved over to the new system.

The MonitorManager::alter_monitor function wasn't used and needs to be
removed as it used the code that was changed.