Enable C++17
Debian 9 reaches its EOL date on June 30, 2022. This means that the next
major MaxScale release is only released on platforms that support C++17.

The main changes that caused conflicts were the addition of new types in
C++17 that some of which were emulated by the bsoncxx library used by the
nosqlprotocol. These are fixed in a separate commit.