Add missing values to non-reserved word list
Not all the values that were added for the XA transaction and KILL query
parsing were reserved keywords.

The addition of the USER keyword to the list causes the KILL query parsing
to not work as expected when the query `KILL QUERY ID USER 'bob'` is
done. Since this is not actually allowed by the MariaDB syntax, rewriting
it to use separate types for the two queries solves both the correctness
problem as well as the parsing problem introduced by the decaying behavior
of USER.

With the new syntax, the ID part in the QUERY ID section is still parsed
as an identifier, not an actual keyword. This is done to avoid adding the
actual keyword ID into sqlite which is already used by it for the
identifier type.