Fix mxs421_events
Now that the configuration is parsed before being uploaded to the MaxScale
VM, the method of un-commenting parameters to enable them is no longer
usable as the comments are stripped out during the parsing process.

To work around this, the sed command can be changed to instead inject the
required values into the configuration. Not the prettiest way to do it but
it works.