Make the causal_reads=global test more consistent
There's no need to use "real" replication lag to test that the feature
works, the MASTER_DELAY is a better alternative as it adds a constant lag
that works on all systems. This makes sure that the test always tests the
parts of the code we want it to test.