MXS-3663: Treat RW and RO transactions similarly
Both read-write and read-only transactions are now handled in the same
way. If the target of the transaction has been selected, it is used as
long as it is valid. If the target is no longer valid, and transaction
replay is enabled, the replay is started.

Since the target selection logic is now bypassed in start_gtid_probe(),
the routing hint is no longer required if the next target to receive it is
a server. However as there can be other services, leaving the hint in the
query should make it more likely that it will end up being routed to an
actual master server.