MXS-3663: Fix handling of empty transactions
A transaction is considered empty if it consists of a single statement. In
practice this happens when a SET autocommit=1 is done while autocommit is
already enabled or when a COMMIT statement is used outside of a
transaction.

The call to finish_transaction() should only be done when a transaction is
not empty. This isn't required by the code, apart from the info level
logging it does, but the logic of the transactions in readwritesplit
assume that non-empty transactions all have a target. Otherwise the debug
assertion will break.