MXS-3663: Retry GTID probing
The retrying of the GTID probe is now tied to delayed_retry and
delayed_retry_timeout. This makes it behave the same way other statements
behave when it is enabled, be it directly or via transaction_replay. The
same function can be used in both the handleError code when the master
where it is executed dies and when the transaction is being replayed due
to it being interrupted.