6.2 Remove check from system-test/xpand_transaction_replay
The test assumed that if an xpand node other than the one we are
connected to is taken down, we will after the replay continue to
be connected to the same node.

That's no longer the case and also not something the test should
be concerned about, so that check was removed.