Fix connection capabilities
The correct way to calculate the final connection capabilities is to AND
the bitmask sent by the client with the one MaxScale sent in the
handshake.

The capabilities of the connection were taken directly from the client's
handshake. This meant that capabilities that MaxScale didn't send in the
handshake were accepted as valid capabilities.

The backend capabilities were always pruned to a bitmask that wasn't the
same as the one that was sent in the handshake. This could cause the
client connection to have different capabilities than the backend
connection.