Amend debug assertion with known problems
When an upper-level service routes a BEGIN statement to one service, all
other services that it also connects to will behave as if they had an open
transaction as well. This is mainly a problem in the debug assertions
which fail when session commands are routed during one of these cases.

To fix this problem, each service would have to have their own version of
the transaction state that is now only stored in the client protocol.