MXS-3642 Improve sidebar navigate method
The config visualization page will utilize nested routes so the
`navigate` method should be handled better.

This adds `redirect` to meta object to check if the route
will be redirected to children route. If page is already at
that redirected route, clicking on that sidebar route won't
perform redirection again.