MXS-3663: Sync all queries routed to slaves
Previously with causal_reads=universal the read was only synchronized if
it was a COM_QUERY with a SELECT inside it. This was different to how the
other statements behaved but that in and of itself wasn't a problem, more
of an inconsistency. The problem is the START TRANSACTION READ ONLY which
would not be synchronized: the GTID at which the transaction would start
would be from before the synchronization and it could cause problems if
used with the snapshot isolation in InnoDB.

The solution to this is to synchronize all statements that actually
execute any SQL, namely COM_QUERY and COM_STMT_EXECUTE, regardless of the
query operation.

The only exception to this are queries of the type QUERY_TYPE_COMMIT or
QUERY_TYPE_ROLLBACK, they never need to be synchronized.