Adjust app padding layout
Horizontal padding of the app layout should be equal to 36px. i.e.
for dashboard page, it is 36px but for details page such as ServerDetail,
ServiceDetail, the left padding is equal to 60px.
This also adds `d-flex flex-column` class names to `page-wrapper`
component to make its children html elements to behave as block element.
By doing this, when assigning `height:100%` or adding `fill-height` class
to one of the children of `page-wrapper` would work without having to subtract
the height of `page-header`.