MXS-3663: Prevent unwanted master replacement
If the target selection for some reason selects a server that isn't
actually master for the target type TARGET_MASTER, it would be
unconditionally used as the new master. Adding a check that the server is
actually labeled as a master prevents this from accidentally happening.

Also extended the info level logging to explain why a query fails to be
routed even if it has a target. Previously the query would silently fail
if the master replacement was not possible.