MXS-3663: Fix GTID probing in transactions
If the master died during the GTID probing inside a transaction, the query
was retried before the transaction replay started. This caused debug
assertions to hit and ended up executing the queries in the wrong order.

The correct thing to do is to only retry the query directly if no
transactions are open.