Fix backend connection capabilities
Using the client capabilities to create the backend capabilities revealed
that the backend protocol sends a very specific type of handshake response
which must enable some capabilities by default.

There are two cases where the capabilities must be derived from the actual
state of the connection: when SSL is enabled for either the listener or
the server but not for both and when the client originally did not connect
with a database but later it did a COM_CHANGE_USER. In the former case the
bits must be disabled so that the server does not expect a SSL request
packet and in the latter case the packet layout changes depending on
whether the capability is enabled or not.