Use atomic operations in WatchdogNotifier
The acquisition of the lock can be replaced by using atomic operations on
the counter. This avoids grabbing the mutex in BackendDCB::connect when
the SystemD watchdog is enabled.

Also removed the redundant notification that was done in the stop()
method. If there is no need to notify anything, there's also no need to
wake up the notifier thread.