MXS-3973: Freeze session capabilities after creation
The capabilities should not change after the session has started. Given
that the filters of a service can be modified at runtime and this triggers
the recalculation of the capabilities, the behavior of the session would
change in ways that it shouldn't.

In practice this is only a problem when a filter is removed from a service
that requires additional capabilities on top of the base service
capabilities. Mainly this is a problem with readconnroute services but it
can also affect readwritesplit if used with the maskingfilter or
binlogfilter.

This also doubles as a tiny performance improvement as it avoid the need
to perform a lookup into the WorkerGlobal for each request and
response. This in turn reduces the number of Worker::get_current() calls
by a large number.