MXS-3961 Enable `forceFallback` option for the drag/drop feat
When dragging a slave server node to a master server node, there is a
need to change the HTML content of the dragging slave server to
inform the user what the operation is.
By default, sortablejs uses HTML5 drag/drop API, so it won't
generate a copy of the drag element. This makes it impossible
to change the HTML content of the drag element.

This enables `forceFallback` to generate a copy of the drag element
having `rect-node-clone` as class name.