MXS-3865: Fix shutdown hang
The MainWorker would run into a deadlock with the libmicrohttpd daemon
thread if the shutdown started at the same time a REST API request was
done. This happens because the daemon thread waits for the MainWorker to
process the request before accepting more. This in turn would cause a
deadlock when the MHD_stop_daemon() call, done in the MainWorker thread,
would try to acquire a lock that was held by the daemon thread.