Fix crash on nested parameter validation
If the module did not provide a specification, it would crash as the
pointer was accessed before it was verified to not be null.

Additionally, modules with specifications but not with prefixes and
modules with no specification at all were not handled the same way:
unknown nested parameters for the former were always treated as valid
whereas they were always invalid for the latter. Both of them should treat
unknown parameters invalid if any were defined.