Don't sort monitor relationships
At least mariadbmon looks at the server order for preference in some
operations. To make it easier to control this at runtime, the
relationships must be treated as ordered.

This also masked a broken test case where the link command would accept a
duplicate of one of the servers. The correct thing is now also done and
the linking of the same server twice is treated as an error.