Fix differences in starting MainWorker
This is really just a band aid for allowing MainWorker to run
via MainWorker::start() in system test. The gateway calls
MainWorker::run() directly, which means the MainWorker is both
created and running in the main thread, so setting the thread
local works correctly when set in the constructor. With start()
it has to be reset in pre_run().