MXS-3961 Extend `details-page-title` to have list item style
When clicking the gear icon, instead of showing a list of icons in
a horizontal style, it should have option to show a list of items or icons
in a vertical style. The `Cluster` page will re-use this component to
show a list of `module_command` to control replication.

This adds `setting-menu-list-item` slot to render items vertically.