MXS-3961 Refactor `rep-tooltip` to make it easier to re-use
The component is used to:
_Show slave replication status of a slave server when hovering on
a slave server.
_Show an overal replication status of all slaves servers when
hovering on a master server.
The component is refactored to accept two required props .i.e
`serverInfo`, an array with one item which holds replication status
info for a slave server if `isMaster` props is false, otherwise
it's an array of multiple items.

This refactors it so that the `cluster-node` component can use
to show replication status.