MXS-3961 Disable `inheritAttrs` to wrap `rep-tooltip`
By default, when `inheritAttrs` is true, parent scope attribute
bindings are not recognized as props and be applied as normal HTML
attributes. So when this is disabled, the attributes are available
via the `$attrs` instance property and can be bound as props by using
`v-bind`. So `rep-tooltip` won't have to declare props passed to
`v-menu`. The only downside is that, if the props is a boolean type,
the attribute passing has to explicitly assign the value.
e.g. Do this `<rep-tooltip :top="true" />` instead of this `<rep-tooltip top />`