MXS-3902 Pool connection only when required
Unnecessary pooling of connections should be avoided. When
deciding whether to pool a connection, check if a session is
waiting for a connection to the server. Only pool if the
connection is required. This change, when used with correct
configuration, greatly reduces unnecessary pooling/reuse of
connections.

ReadWriteSplit can route reads even without a master connection.
To benefit from this, the pooling logic needs to tell the
router when a connection is unusable. The exact logic if any is
router-specific.

In a typical situation, routing connections to slave servers are
spread out to multiple servers. Likely, such connections do not
need to be pooled/multiplexed as much as connections to
the master.