MXS-3951: Add ParamServerList
This type will be needed by mariadbmon for the servers_no_promotion
parameter.

The unit test also had a problem where the servers were correctly deleted
by the std::unique_ptr but a stale reference was still stored in the
actual ServerManager code. The correct way to handle this is to call
ServerManager::destroy_all once the test is complete.

Using this parameter type for services and monitors is more complex: the
parameters themselves aren't modified by the REST API and the object
relationships are defined by the relationships JSON in the payload.