MXS-3951: Configure monitors using Configuration
The monitors are now created and altered using the native types:
ConfigParameters for the static files and JSON for the runtime changes.

To make the act of serializing the server relationship easier and to make
the configuration of servers behave the same way for both statically and
dynamically created objects, the servers parameter is now always injected
into the configuration where the monitor is constructed or altered from.