MXS-3961 Use new UI for showing replication stats instead of using tooltip
Showing replication status in a tooltip is clumsy in this case.
From the UX perspective, showing a tooltip every time user hovers
on the card is not intuitive. e.g. If user just want to mouseover
the gear icon or drag&drop the card, the tooltip still shows.
Even adding the info icon to resolve this, it's still clumsy due to the
current card layout.

With new UI, there is an arrow button below the card, clicking it
will increase the height of the card to show "extra" info. This
extra info section can be then a carousel with different info slides.