MXS-3961 Handle dynamic height for nodeSize
Cluster node's height during expanded transition time is
determined by counting the number of new lines are added when the user
clicks the expand arrow button. Though it's possible to get the height
of the cluster node card after expanding by adding a delay equal to
the duration of the transition, this causes a bad transition. Because
the graph will redraw only after the cluster node is fully expanded which
is a clumsy transition.