MXS-3951: Inject servers into the correct parameters
Previously they were inserted into the new JSON payload which wasn't
actually the one used with the reconfiguration of the monitors. This
caused all attempts to alter the servers of a monitor to fail.

The creation of the parameter object was also not needed as at the
point the function is called, the final parameters object always
existed.