MXS-3961 Replace `getNodeAttrs` with `getServerData`
Instead of manually destructuring properties to shown and spreading
it to each cluster node, this assigns server's data to `serverData`
property. By doing this, the `cluster-node` component can then extract
any properties it needs.

This also adds a check in the `cluster` view to check if the `cluster-node`
component is passed with `node` object having `data` object as property.
This is because `data` object may be undefined when `tree-graph` is still
rendering the node. The helps to remove unnecessary type checking in
`cluster-node`. Type checking should only be used when there is a possibility
that the property in the `node` object could be undefined.