MXS-3951: Fix monitor relationship handling
As the monitors depend heavily on the base configuration of the
mxs::Monitor class, a change in it must also trigger the
post_configure() of the module's configuration to be called. This must
be done as the monitors have internal bookkeeping that refers to
servers derived from the base class's configuration which must be
reset whenever it changes (which is pretty much always).

In addition, the monitor relationships should only be modified at
runtime if an actual JSON object is defined for it.