MXS-3951: Fix MaxCtrl unit tests
The new system always adds the unit suffix to the returned value which
caused the test to fail.

In addition, the real value used is now returned which for the
timeouts means that a different value is returned than they are
configured with. This only happens if milliseconds are used to
configure a parameter whose granularity is one second.