Fix run_ctrl_c.sh
If the script is run on the CentOS 7 VM which has the newer compiler
from the devtoolset, the sudo command in it is just a wrapper around
"scl enable". This means that the ASAN options were interpreted as a
command to execute instead of environment variables like they normally
would.

Since the devtoolset isn't relevant for this test, we can simply use
the absolute path to the real sudo executable to work around this.