Check for expired connections once per second
This keeps the true granularity of persistmaxtime closer to the
documented one second granularity and also fixes the pers_01 test
which would fail as the connections wouldn't be removed when it
expected them to be.

The real cost of doing this check once per second instead of once
every five seconds should be evaluated as the cost of increasing the
interval reduces the accuracy of persistpoolmax which has to be
documented. Ideally non modulo-5 values would also be rejected.