MXS-3961 Place the `circle` node in a better position in the graph
The `circle` node was place on the left side of the rectangular box,
so the box always covers the connection link between nodes which makes
it difficult to see the connection between nodes.

This places the `circle` node on the right side of the box if it
has children nodes. This makes it easier to recognize that the circle
node can be expanded/collapsed.