MXS-3961 Handle either showing address:port or socket
Combining address:port or socket with the server name in the title
section makes it look messy. It should show only the server name
in the title section. The `address:port` or `socket` will be shown
in the body of the `cluster-node`.

This also reduces carousel's delimiter control height to 20px.