Use UPDATE in mxs2512_trx_replay_rollback
SELECT ... FOR UPDATE causing a deadlock cannot be used anymore now that
MDEV-27025 was fixed in 10.6.6. Using an UPDATE is a better method as it
causes an actual conflict to take place.

Also removed the unused Result type and other related code and switched
over to using the TestConnections::expect() instead of assertions. This
prevents the test cleanup from being skipped if the test failed.