MXS-4001 Check also Redis error when checking connection state
When the Redis client has connected to the Redis server, a
CONNECTED bit will be set. However, if Redis closes the connection,
the bit will not be cleared, but a separate error will be set. Thus,
when checking whether the connection is usable, one has to check
both the CONNECTED bit and the error state.

If Redis has been configured with a timeout, then it will close the
connection has been idle. That must be detected, so that the connection
is re-created.