MXS-3961 Resolve `/query/:id` nested route
The query page can be now accessed with an absolute link. If the param
id (the resource name) matching with a connection name which is already
bound to a worksheet, it chooses that worksheet as the current active
worksheet. Otherwise, it creates new blank worksheet and then automatically
opens the connection dialog with pre-select data.

The opening of the connection dialog is not yet implemented in this commit.