MXS-4006 Add INTERPRET_AS_SECONDS_RELAXED
Differs from INTERPRET_AS_SECONDS in that a non-suffixed
number is interpreted as seconds, yet allows durations
that are not multiples of a second including durations that
are less than a second.

This is for the pooling mechanism. In 7 duration-suffixes
will be required, which removes the need for interpretation
altogether.