Move core luafilter functionality into LuaContext
Helps avoid the duplication that comes with the session and global
scripts doing the same thing.

This commit also simplifies the global state management by passing the
buffers as arguments instead of using global variables for it. This
prevents the race condition that was present in the old code where
multiple luafilter instances would access the same variable at the same
time.