Expose the session to the Lua functions
This exposes the client's user, host and current database to the Lua
side of things. Replaced the ID counter function with a function that
returns the current session ID.

Added the name of the server where the response came from to the
clientReply API function. This makes it possible to implement
rudimentary auditing of where queries were executed.

Renamed the functions to start with a `mxs` prefix to make them stand
out more. Since the filter is experimental, we don't have to care about
backwards compatibility that much.