MXS-4004: Delay the OK packet for KILL commands
Sending the OK packet after all the backend connections have finished
executing the KILL command will cause the client to send the COM_QUIT at
the right time. Previously it was fairly common for the client
connection to be close before the backend connections were open and the
KILL commands were executed. This is also functionally closer to how the
actual KILL command is handled in MariaDB.

The old code used a detached thread to collect the information from all
the RoutingWorkers but this could in theory cause problems if the thread
was too slow and only started up after the shutdown process had already
started. By executing the code on the MainWorker, we avoid this
problem. This is not an optimal solution as it places unwanted stress on
the MainWorker which could be busy doing something else, possibly
something that takes several seconds. A better, albeit more complex,
solution would be to broadcast the request for information to all
RoutingWorkers and then once the last of them has collected it, it would
move the execution back to the RoutingWorker that owns the session.

This also solves MXS-4002 by adding info level log messages for the KILL
command processing.