Type check Lua globals before calling them
This makes sure that the global variables are functions and not
something else before we attempt to call them.

Moving the code into a generic function also makes the error reporting
more consistent across all the API functions. Previously the error
messages were different based on which function was called.

This commit changes the semantics of the routeQuery API function so that
it is called once for every routed packet instead of every text-form SQL
query. This makes it so that there is always at least one routeQuery
call before the clientReply callis made.