MXS-4004: Don't send OK with mxs_mysql_execute_kill
The function is used by the smartrouter which needs it to kill all the
active queries. When called via mxs_mysql_execute_kill, the OK packet
must not be sent as it generates an extra response the client does not
expect.

This fixes the sr_basic test which previously failed with a debug
assertion as was reported in MXS-4016.