MXS-3945: Fix reading of binary mode Avro data
The binary mode modified the file position to read the raw data but it
did not adjust it back to the end of the block after the read was
done. This caused the block sync marker to be interpreted as the record
count and block size which caused all sorts of problems as the data was
essentially garbage.