MXS-4030 Fix `numericFields` in `visualize-sidebar`
When visualizing the query result, the y axis input dropdown fails
to show some columns if the cell value is a number string. e.g. "10.5".
It fails to recognize that value is a numeric.

This adds `IsNumericCell` method to convert it to a number using
`parseFloat` before checking its type.