MXS-3808: Don't flush pool for REST API requests
This is very bad for the REST API performance as it causes the workers
to iterate over all the pooled connections. In addition to this, under
heavy workloads the workers might be slow to respond which further slows
down the requests. Since the only reason why it was done in the first
place was to make tests pass, an alternative for it can be added later
on.