MXS-3993: Use real server capabilities
The set of capabilities that MaxScale now sends is the subset of
capabilities that all the backend servers support. This allows the
backend server to fake its version without it affecting the capabilities
that MaxScale sends. However, in practice this only happens if an older
MariaDB version claims to be newer than it is.