MXS-3808: Access pool stats directly
Under heavy loads the time it takes for the workers to answer the
request for pool statistics can be comparatively long. In addition to
the RoutingWorker, the only other user of the pool statistics is the
REST API. Protecting the access to the statistics with a mutex allows
the REST API to access the data directly without having to wait for the
worker to finish processing the current batch of events.

In 2.5 this was not a problem as all the RoutingWorkers updated the same
set of statistics but in 6.2 each worker has their own copy.