MXS-3961 Declare portal to `page-header--right` in each page
`page-header--right` was declared in `details-page-title` to dry code
but it brings complexity to `create-resource` component in terms of
handle showing desired form type. Currently `create-resource` component
uses `$route.name` to choose type of form. e.g. If the `create-resource`
dialog is opened in `/dashboard/monitors/Monitor` route, it chooses
the monitor form to show which is not what the user always wants.
Besides, there is `form_type` state to override this behavior which
brings an additional complexity into the component. Now, there is a
need to use `create-resource` component in the cluster page, it would
be difficult to hanlde this as the cluster route has `cluster` as
the route name.

To eliminate the complexity, the `create-resource` component should
replace those handlers with an interceptor to choose form type before
dialog is opened.

This commit is a first step of achieving that by declaring `create-resource`
component in all needed pages.