MXS-3961 Refactor `create-resource` to accept default form type
Clicking the `+ Create New` button should select the default form
depending on what page it is. If the user is on the monitor, service
detail or cluster page, it selects server form. Otherwise, service form.

This adds `defFormType` props which allows selecting default form
after dialog is opened. `form_type` is kept for an exceptional case
where the dialog should be opened without clicking the `+ Create New`
button.

This also cleans up the component by removing unnecessary and obsolete
codes.