MXS-4034: Periodically clean up the connection pool
This makes the persistent connection pool in 2.5 behave like it does in
newer versions. The implementation is still overall inefficient as the
pool is purged right before a DCB is taken from it. The code could just
find the first suitable DCB and leave the cleanup to be done as a batch
process.