MXS-3961 Replace `$route.name` with `defRelationshipType` props in `create-resource`
Relying on `$route.name` narrows down the ability of choosing default
relationship value for the object being created. e.g. The route name
of the cluster page is `cluster` which is a synonym of `monitor`. So
using `$route.name` won't work.

This replaces it with `defRelationshipObj` props. This props uses to
select the default relationship object.