MXS-4038: Ignore rate limit for `reload service`
The UserAccountManager now always does the update after a call to
update_user_accounts() is done. This allows the calling code to decide
whether to respect the rate limit or not. As the limit is now opt-in the
`reload service` command should now work.

This fixes the mxs922_bad_server test which relied on the users being
updated once the configuration was fixed. Previously the test would fail
when the users were not updated in time.