MXS-3931: Build cert chain during configuration
This allows non-fatal errors to be logged during configuration if any
are found. This is also a performance improvement as otherwise the
certificate chain would be built separately for every time a new SSL
object is created.

If the file given to ssl_cert does not contain the full certificate
chain, OpenSSL will report it as a problem. Since it's not a functional
problem if file doesn't contain the full chain, the messages are logged
on the notice level and an extra message is added to explain it.