MXS-4035 Report failure to parse as info-message
If invalidation is enabled and the parsing of a statement fails,
then the resultset cannot be cached, as we would not know when to
invalidate it (since we don't know what tables it accesses).

However, since this only affects the performance and not the
functionalty, and since if there is one such statement it will
be warned for over and over again, it is better to log this as an
info-level message.