MXS-3990 Do not rely upon rworker ids starting from 0
As the worked id generation will be moved to mxb::Worker, it can
no longer be assumed that the ids of the routing workers start
from 0, but the code must work regardled of the range of the
routing worker ids. However, the code does assume that all routing
workers get consequtive ids.