MXS-3990 Make the delayed call id globally unique
The 16 most significant bits of the 64 bit delayed call id,
are the 16 least significant bits of the worker id. So, as
long as there are no more than 65536 workers we are fine.

With this arrangement, the delayed call id will be globally
unique, which makes moving a pending delayed call from one
worker to another simpler.