MXS-3990 Prepare for delayed_call life-time management
A delayed_call can now only be made in the context of a Worker::Object
object. That will be used for life-time management, as it must be ensured
that a delayed call is not made on an object that no longer exists.

MXS_SESSION is now a Worker::Object and exposed delayed_call functions
that should be used whenever the delayed_call takes place in the context
of a session.

In subsequent commits, the actual life-time management will be added.